+++ CHANGELOG.md
@@ -4,8 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [Unreleased](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v14.2.1...HEAD)
+# [Unreleased](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v14.2.2...HEAD)
+
+# [v14.2.2](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v14.2.1...v14.2.2)
 * [PCM-1993](https://inindca.atlassian.net/browse/PCM-1993) - Handled `reject` messages and treat the corresponding session as handled
+
 
 # [v14.2.1](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v14.2.0...v14.2.1)
 * [PCM-1988](https://inindca.atlassian.net/browse/PCM-1988) – Add jidResource support for jwt tokens.